refactor(archive): 下沉标题归一化/匹配/批量过滤到 bangumi_archive 模块
_archive_shortcut.py 与 bangumi_archive/ 职责重叠：
- 标题归一化工具（_strip_season_episode_suffix / _strip_title_wrappers /
  _split_title_segments + 4 组常量）已被 bangumi_api/search.py 跨模块导入，
  本应属于 archive 模块而非 shortcut
- _find_subject_ids_for_title / _try_main_segment_match 依赖
  archive_title_index，是标题查询能力的扩展
- _fetch_and_filter_subjects 依赖 archive_store.get_subject，与 store 的
  查询职责重叠（store 已有 get_subject(subject_type=) 过滤能力）

下沉变更：
- 新建 bangumi_archive/_title_normalize.py：4 组常量 + 3 个归一化函数
- archive_store 新增 get_subjects_by_ids_with_filter（批量拉取+过滤）+
  MAX_IDS_TO_FETCH 类属性（原 _MAX_IDS_TO_FETCH）
- archive_title_index 新增 find_subject_ids_for_query_title（多策略标题
  匹配）+ _try_main_segment_match

简化效果：
- _archive_shortcut.py 从 907 行降至 530 行（-42%），仅保留 ShortcutResult
  + ArchiveShortcut 的 try_* 命中判断薄壳
- search.py 导入路径改为从 bangumi_archive/_title_normalize 导入
- 测试 mock 路径从源模块改为导入位置（_archive_shortcut.archive_title_index）

全量 3074 测试通过。
